Resolve the SDDM greeter's keyboard layout from vconsole.conf
SDDM launches the greeter's compositor with an explicit --config, so nothing
under default/hypr is loaded and Hyprland falls back to its built-in "us"
layout. The greeter is the only place a password is typed by hand, so on any
non-US layout it rejects correct passwords with no indication why. Autologin
hides this until the first manual logout.

Move the vconsole.conf resolution out of default/hypr/input.lua into a shared
default/hypr/keyboard.lua and import it from both the session config and the
greeter, so a password types identically in both. The greeter loads it with
dofile because its environment sets neither package.path nor $HOME.
